// src/app/views/home/enums/enum_home.ts
export enum HomeEnum {
    TABLE_LISTAS_PRIMARYKEY = 'listId',
    TABLE_LISTAS_NAME = 'listas',
    TABLE_TARJETAS_PRIMARYKEY = 'cardId',
    TABLE_TARJETAS_NAME = 'tarjetas',
    CARD_EDITING_DELAY = 300,
    ID_INPUT_EDIT_LIST = 'input-edit-list-',
    ID_INPUT_EDIT_CARD = 'input-edit-card-',
    ID_TEXTAREA_CARD = 'textarea-card',
}